Run3Detector/analysis/pulseInjection/GAN/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
awkward==2.6.6
awkward-cpp==35
certifi==2024.7.4
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.1
h5py==3.11.0
idna==3.7
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
-e git+ssh://git@github.com/milliQan-sw/milliqanOffline.git@7c90c916594f4d6985d19f3d3ec882a0b9258524#egg=local_gan&subdirectory=Run3Detector/analysis/pulseInjection/GAN
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
protobuf==4.25.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.7.1
scipy==1.14.0
setuptools==71.1.0
six==1.16.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-probability==0.24.0
termcolor==2.4.0
tf_keras==2.17.0
typing_extensions==4.12.2
uproot==5.3.10
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
